fix(clippy): close the last three findings, so the Lints gate can be green
`Lints` is fmt, then `cargo clippy --tests --benches --examples -- -D warnings`.
Formatting was one half of why it has been red on main; these three are the
other. All are in files this branch does not otherwise touch, and all are the
mechanical kind — the sort a gate is supposed to catch on the commit that
introduces them, not accumulate.

- dns_finality.rs: `o.value as u64` where `value` is already u64. A cast that
  does nothing still tells the next reader the types differ.
- palw_economic_locus_v1.rs: `A % B == 0` is `A.is_multiple_of(B)`, which is what
  the assertion's own message ("a whole number of drains") already says.
- palw_state_v2.rs: `&vec![8u8; 4]` allocated a Vec to borrow it as a slice.

`cargo clippy --tests --benches --examples -- -D warnings` now exits 0, `cargo
fmt --all -- --check` is clean, and kaspa-consensus-core's 1,940 lib tests pass —
two of the three edits are inside assertions, so the suite is the check that they
still assert the same thing.

rust-toolchain.toml records eleven outstanding findings against 1.93.0 as
"outstanding work, not an accepted baseline". Three remained; there are none now.

+++ consensus/core/src/dns_finality.rs
@@ -11903,7 +11903,7 @@ mod tests {
             "past the fence one block releases at most one cap however many thresholds it crossed ({bounded_total} > {cap})"
         );
         assert!(
-            bounded.iter().map(|o| o.value as u64).sum::<u64>() == bounded_total,
+            bounded.iter().map(|o| o.value).sum::<u64>() == bounded_total,
             "the minted total and the outputs are one statement — the reserve recurrence subtracts exactly what was paid"
         );
